Changed paths: M       CHANGES.txt M       cat_main.m M       cat_surf_display.m M       cat_surf_render.m
(1) Fixed: Wrong weighting of NCstr in cat_main.m
(2) Changed: Use try for call of cat_surf_display and cat_surf_render because of
    weird error in Matlkab 2015a with non-double values for Faces.


git-svn-id: svn://141.35.69.214/cat12@884 9f44b166-36c4-483c-8295-4af200601734

+++ cat_surf_render.m
@@ -88,7 +88,7 @@ switch lower(action)
             for mi = 2:numel(MS)
                 try
                     MI         = gifti(MS{mi});
-                    M.faces    = [M.faces; MI.faces + size(M.vertices,1)];   % further faces with increased vertices ids
+                    M.faces    = [double(M.faces); double(MI.faces) + size(M.vertices,1)];   % further faces with increased vertices ids
                     M.vertices = [M.vertices; MI.vertices];                 % further points at the end of the list
                     if isfield(M,'cdata');
                         M.cdata  = [M.cdata; MI.cdata];                     % further texture values at the end of the list
@@ -242,7 +242,7 @@ switch lower(action)
         
         %-Patch
         %------------------------------------------------------------------
-        P = struct('vertices',M.vertices, 'faces',M.faces);
+        P = struct('vertices',M.vertices, 'faces',double(M.faces));
         H.patch = patch(P,...
             'FaceColor',        [0.6 0.6 0.6],...
             'EdgeColor',        'none',...
